Finished proyect

+++ Angular7/src/app/employees/employee-list/employee-list.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { EmployeeService } from 'src/app/shared/employee.service';
 import { Employee } from 'src/app/shared/employee.model';
 import { ToastrService } from 'ngx-toastr';
+import { IfStmt } from '@angular/compiler';
 @Component({
   selector: 'app-employee-list',
   templateUrl: './employee-list.component.html',
@@ -22,9 +23,13 @@ export class EmployeeListComponent implements OnInit {
   }
   onDelete(id:number)
   {
-    this.service.deleteEmployee(id).subscribe(res=>
-      {
-          this.service.refreshList();
-      });
+    if(confirm("Are you realy sure of delete this record?"))
+    {
+      this.service.deleteEmployee(id).subscribe(res=>
+        {
+            this.service.refreshList();
+            this.toastr.warning("Deleted succesfully","EMP. Register")
+        });
+    }
   }
 }
